feat: move form logic to hook #84
