+++ .eslintrc
@@ -19,6 +19,6 @@
     "no-console": "warn",
     "no-eval": "error",
     "import/first": "error",
-    "react/prop-types": 0,
+    "react/prop-types": 0
   }
 }
