+++ .prettierrc
@@ -1,4 +1,4 @@
 {
 	semi: false,
-	printWidth: 75,
+	printWidth: 75
 }
