+++ package.json
@@ -1,8 +1,15 @@
 {
   "name": "payp",
   "version": "0.1.0",
-  "private": true,
-  "homepage": "http://bravedevs.epizy.com/",
+  "homepage": "http://payp.epizy.com/",
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
+    "lint": "eslint \"src/**/*.{js,jsx}\"",
+    "format": "prettier --write \"src/**/*+(.js|.css|.jsx)\""
+  },
   "dependencies": {
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
@@ -16,17 +23,6 @@
     "screen-cover": "^0.1.5",
     "styled-components": "^4.4.0",
     "tachyons": "^4.11.1"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "lint": "eslint src/**/*.{js,jsx}",
-    "format": "prettier --write \"src/**/*+(.js|.css|.jsx)\""
-  },
-  "eslintConfig": {
-    "extends": "react-app"
   },
   "browserslist": {
     "production": [
